feat: refactor of ui to the correct folder
