+++ index.html
@@ -16,7 +16,7 @@
         
     </div>
     
-    <script type="module" src="./src/app/index.tsx"></script>
+    <script type="module" src="./src/ui/index.tsx"></script>
     <script type="module" src="./src/main.ts"></script>
 </body>
