--- src/app/components/GameCanvas/GameCanvas.tsx
@@ -1,11 +1,0 @@
-import React, { useRef } from 'react';
-
-const GameCanvas = () => {
-    const ref = useRef<HTMLCanvasElement>(null);
-    
-    return (
-        <canvas id="game"></canvas>
-    );
-}
-
-export default GameCanvas;
