+++ src/ui/components/PlayerDeadPopup/PlayerDeadPopup.tsx
@@ -1,5 +1,5 @@
 import { useCallback, useEffect, useState } from "react";
-import usePlayerInfo from "app/clientState/providers/PlayerInfoProvider/usePlayerInfo";
+import usePlayerInfo from "ui/clientState/providers/PlayerInfoProvider/usePlayerInfo";
 
 const PlayerDeadPopup = () => {
     const { isDead, isDeadPopupCallback } = usePlayerInfo();
