+++ src/ui/components/PlayerInventory/InventoryItemList.tsx
