+++ src/ui/components/PlayerInventory/contract.ts
