+++ src/ui/index.tsx
@@ -1,5 +1,5 @@
 import { createRoot } from "react-dom/client";
-import { App } from "./App";
+import { UI } from "./UI";
 
 const container = document.getElementById("ui");
 
@@ -8,4 +8,4 @@ if (!container) {
 }
 
 const root = createRoot(container);
-root.render(<App />);
+root.render(<UI />);
