+++ src/ui/utils/pixelRatioUtils.ts
@@ -1,3 +1,13 @@
+import { Vector } from "excalibur";
+
+export const calculateExPixelConversion = (screen: ex.Screen) => {
+  const origin = screen.worldToPageCoordinates(Vector.Zero);
+  const singlePixel = screen.worldToPageCoordinates(new Vector(1, 0)).sub(origin);
+  const pixelConversion = singlePixel.x;
+  document.documentElement.style.setProperty('--pixel-conversion', pixelConversion.toString());
+}
+
+
 export const getPixelRatio = () => {
     return parseFloat(getComputedStyle(document.documentElement).getPropertyValue('--pixel-conversion')) || 1;
 };
